Isolate common types for AstarteVoyagerIngress
